Move command validation logic
